+++ dmr/_compiled/negotiation.py
@@ -1,27 +1,6 @@
 # Some code here was adapted from Litestar under MIT license
 # https://github.com/litestar-org/litestar/blob/main/LICENSE
 
-# The MIT License (MIT)
-#
-# Copyright (c) 2021, 2022, 2023, 2024, 2025, 2026 Litestar Org.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
 import re
 from collections.abc import Iterable
 from typing import Final, final
@@ -99,20 +78,16 @@ def accepted_header(accept_value: str, media_type: str) -> bool:  # noqa: C901
 
     For example:
 
-        .. code:: python
+    .. code:: python
 
-            >>> from django.http import HttpRequest
-            >>> request = HttpRequest()
-            >>> request.META = {'HTTP_ACCEPT': 'application/json'}
-            >>> accepted_header(
-            ...     request.headers.get('Accept'), 'text/plain'
-            ... )  # equivalent to request.accepts("text/plain")
-            False
-            >>> accepted_header(
-            ...     'application/json,text/html;q=0.8',
-            ...     'application/json',
-            ... )  # can be called with any headers-like mapping
-            True
+        >>> assert accepted_header('application/json', 'text/plain') is False
+        >>> assert (
+        ...     accepted_header(
+        ...         'application/json,text/html;q=0.8',
+        ...         'application/json',
+        ...     )
+        ...     is True
+        ... )
 
     """
     if not accept_value or not media_type: